Move preview controller logic into concern
